chore: resort test code
